fix: split user requests handler into create user request handler and get users request handler.
