+++ src/server/handlers/index.js
@@ -1,2 +1,3 @@
-export * from "./user-requests-handler.js";
+export * from "./create-user-requests-handler.js";
+export * from "./get-users-request-handler.js";
 export * from "./not-existing-requests-handler.js";
